+++ coop/q_shared.c
@@ -257,7 +257,7 @@ float Q_fabs (float f)
 
 #if defined(_MSC_VER) && defined(_M_IX86) && !defined(C_ONLY)
 #pragma warning (disable:4035)
-__declspec( naked ) long Q_ftol( float f )
+__declspec( naked ) int Q_ftol( float f )
 {
 	static int tmp;
 	__asm fld dword ptr [esp+4]
