+++ qcommon/qcommon.h
@@ -780,7 +780,7 @@ MISC
 void		Com_BeginRedirect (int target, char *buffer, int buffersize, void (*flush));
 void		Com_EndRedirect (void);
 void 		Com_Printf (char *fmt, ...) __attribute__((__format__(__printf__,1,2)));
-void 		Com_DPrintf (unsigned long developerFlags, char *fmt, ...) __attribute__((__format__(__printf__,2,3))); /* FS: Added developer flags */
+void 		Com_DPrintf (unsigned int developerFlags, char *fmt, ...) __attribute__((__format__(__printf__,2,3))); /* FS: Added developer flags */
 void 		Com_Error (int code, char *fmt, ...) __attribute__((__noreturn__, __format__(__printf__,2,3)));
 void 		Com_Quit (void) __attribute__((__noreturn__));
